[Add/Remove] Removed some previous attempts at this function, and created a working version in models/playlists
